fix(discord): keep task card count bigint-safe (#4055)

+++ src/services/discord/task_notification_delivery/mod.rs
@@ -242,7 +242,7 @@ enum TaskCardPayload {
 }
 
 impl TaskCardPayload {
-    fn render(&self, update_count: u32) -> String {
+    fn render(&self, update_count: u64) -> String {
         match self {
             Self::Task(note) => {
                 super::tui_task_card::format_task_notification_card(note, update_count)
